requirements.txt
# ============================================================
# Azure SDK
# ============================================================
azure-storage-blob==12.19.1
azure-identity==1.15.0

# ============================================================
# Azure OpenAI
# ============================================================
openai==1.14.3
httpx==0.27.0

# ============================================================
# PDF extraction
# ============================================================
PyMuPDF==1.24.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# ============================================================
# Chunking / LlamaIndex
# ============================================================
llama-index-core==0.10.30
llama-index-embeddings-openai==0.1.9
llama-index-embeddings-huggingface==0.2.0

# ============================================================
# Embeddings (local fallback)
# ============================================================
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.39.3

# ============================================================
# Vector store
# ============================================================
chromadb==0.4.24

# ============================================================
# BM25 / keyword search
# ============================================================
rank-bm25==0.2.2

# ============================================================
# Reranking
# ============================================================
# cross-encoder is part of sentence-transformers (above)

# ============================================================
# Markdown parsing
# ============================================================
markdown==3.6
beautifulsoup4==4.12.3

# ============================================================
# Utilities
# ============================================================
python-dotenv==1.0.1
tqdm==4.66.2
numpy==1.26.4
pydantic==2.6.4

# ============================================================
# Notebook
# ============================================================
jupyter==1.0.0
ipywidgets==8.1.2
matplotlib==3.8.4
pandas==2.2.1
seaborn==0.13.2
